backend/requirements.txt
# === Runtime ===
fastapi==0.111.0
uvicorn[standard]==0.30.1
pydantic==2.7.4
pydantic-core==2.18.4
pydantic-settings==2.3.4
sqlalchemy==2.0.32
alembic==1.13.2
psycopg[binary]==3.1.19
httpx==0.27.0
python-multipart==0.0.9
tenacity==8.4.2
loguru==0.7.2
google-generativeai
transformers==4.44.2
torch>=2.2.0
pillow>=10.3
httpx>=0.27
joblib
pandas
lightgbm
langgraph

httpx[http2] 
sentinelhub 
numpy 
python-dotenv

# LangChain + OpenAI 
langchain
langchain-core
langchain-community
langchain-openai
openai
langchain-chroma
chromadb

# === Testing ===
pytest==8.2.0
